fix(blob): extract padding chars from hashes
I didn't previously understand that the "==" at the end of the hash
isn't merely a delimiter, but a padding character for base64 encoding.
It's not strictly necessary to have it, but you have to add it back in
to perform a base64 decode. Base64 encoded strings are padded out to
a multiple of four characters.